Update the XML application profile parser
Currently one problem of KeyMinder is the application specific "SSH port
forwarding". To unify this for all programs, the XML application profile
parser had to be updated.

Therefore the structure of XML application profiles has been changed, this
was necessary, because the old syntax wasn’t flexible enough to append
this feature in an acceptable way.